+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.1.59] - 2026-05-10
+- swipe both ways
+
+## [0.1.58] - 2026-05-03
+- export optimisation
+- exporting tasks and settings now possible
+- exporting and importing now moved to settings
+
 ## [0.1.57] - 2026-05-03
 - added time of day heatmap to stats
 - automate screenshot changelog updates on push to `dev`, `staging`, and `main`